Add Staff card is now its own importable component.
